addressing changes and doing cleanups
