+++ frontend/web/components/EditPermissions.tsx
@@ -22,7 +22,7 @@ import Switch from './Switch'
 import TabItem from './base/forms/TabItem'
 import Tabs from './base/forms/Tabs'
 import UserGroupList from './UserGroupList'
-import { PermissionLevel, Req } from 'common/types/requests'
+import { PermissionLevel, Req, PermissionRoleType } from 'common/types/requests'
 import { useGetAvailablePermissionsQuery } from 'common/services/useAvailablePermissions'
 import ConfigProvider from 'common/providers/ConfigProvider'
 import Icon from './Icon'
@@ -576,7 +576,7 @@ const _EditPermissionsModal: FC<EditPermissionModalType> = withAdminPermissions(
 
     const selectPermissions = (
       key: string,
-      value: 'GRANTED' | 'GRANTED_FOR_TAGS' | 'NONE',
+      value: PermissionRoleType,
       tags: number[] = [],
     ) => {
       const updatedPermissions = [
@@ -587,7 +587,7 @@ const _EditPermissionsModal: FC<EditPermissionModalType> = withAdminPermissions(
       const updatedLimitedPermissions = limitedPermissions.filter(
         (v) => v !== key,
       )
-      if (value === 'NONE') {
+      if (value === PermissionRoleType.NONE) {
         setEntityPermissions({
           ...entityPermissions,
           permissions: updatedPermissions,
@@ -603,7 +603,7 @@ const _EditPermissionsModal: FC<EditPermissionModalType> = withAdminPermissions(
           ]),
         })
       }
-      if (value === 'GRANTED_FOR_TAGS') {
+      if (value === PermissionRoleType.GRANTED_FOR_TAGS) {
         setLimitedPermissions(updatedLimitedPermissions.concat([key]))
       } else {
         setLimitedPermissions(updatedLimitedPermissions)
